pyproject.toml
[tool.poetry]
name = "scoach"
version = "0.1.1"
description = "Setup for training Tensorflow models on SLURM clusters."
authors = ["Gabriel Gazola Milan <gabriel.gazola@poli.ufrj.br>"]
license = "GPL-3.0"
readme = "README.md"
homepage = "https://github.com/gabriel-milan/scoach"
repository = "https://github.com/gabriel-milan/scoach"
keywords = ["tensorflow", "machine learning", "hpc", "slurm"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
]
include = [
    "LICENSE"
]

[tool.poetry.dependencies]
python = "^3.8"
minio = "^7.0.0"
dask_jobqueue = "^0.7.0"
Jinja2 = "^3.0.0"
prefect = "^0.15.0"
tensorflow = "^2.0.0"
loguru = "^0.5.0"
Django = "^3.2.0"
PyYAML = "^5.4.0"
typer = "^0.4.0"
psycopg2-binary = "^2.9.0"

[tool.poetry.dev-dependencies]
pytest-cov = "^3.0.0"

[tool.poetry.scripts]
scoach = 'scoach.cli:app'

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
